simulator/requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn==0.31.0
